falta agregar funcion id a caja , modificar de producto y movimiento no anda aun

+++ Ferreteria.Entidades/Producto.cs
@@ -40,7 +40,7 @@
         
         {
             get { return precioProducto;}
-            set { precioProducto = value; } 
+            set { precioProducto = (int)value; } 
         
         }
         //public bool ExisteProducto
